refactor: separate coordinate class from node class
